add MainLogic class as the entry point
